+++ search/src/test/java/net/chesstango/search/smart/alphabeta/MinMaxPruningBlackTest.java
@@ -61,7 +61,7 @@ public class MinMaxPruningBlackTest {
         Game rootGame = setupGame(Color.BLACK, GameStatus.NO_CHECK);
 
         Game childGame = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1);
+        when(quiescence.maximize(eq(childGame), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1L);
 
         Move move = mock(Move.class);
 
@@ -91,10 +91,10 @@ public class MinMaxPruningBlackTest {
         Game rootGame = setupGame(Color.BLACK, GameStatus.NO_CHECK);
 
         Game childGame1 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1);
+        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1L);
 
         Game childGame2 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn(2);
+        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn(2L);
 
         Move move1 = mock(Move.class);
         Move move2 = mock(Move.class);
@@ -129,10 +129,10 @@ public class MinMaxPruningBlackTest {
         Game rootGame = setupGame(Color.BLACK, GameStatus.NO_CHECK);
 
         Game childGame1 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1);
+        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1L);
 
         Game childGame2 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn(GameEvaluator.BLACK_WON);
+        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn((long)GameEvaluator.BLACK_WON);
 
         // childGame3 no llega a evaluarse, dado que existe CuteOff por el mate que se encuentra en childGame2
         Game childGame3 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
@@ -178,13 +178,13 @@ public class MinMaxPruningBlackTest {
         Game rootGame = setupGame(Color.BLACK, GameStatus.NO_CHECK);
 
         Game childGame1 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(GameEvaluator.BLACK_LOST);
+        when(quiescence.maximize(eq(childGame1), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn((long)GameEvaluator.BLACK_LOST);
 
         Game childGame2 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(GameEvaluator.BLACK_LOST);
+        when(quiescence.maximize(eq(childGame2), eq(1), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn((long)GameEvaluator.BLACK_LOST);
 
         Game childGame3 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame3), eq(1),eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(GameEvaluator.BLACK_LOST);
+        when(quiescence.maximize(eq(childGame3), eq(1),eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn((long)GameEvaluator.BLACK_LOST);
 
         Move move1 = moveFactoryBlack.createCaptureMove(PiecePositioned.getPiecePositioned(Square.c3, Piece.KNIGHT_BLACK), PiecePositioned.getPosition(Square.d4));
         Move move2 = moveFactoryBlack.createCaptureMove(PiecePositioned.getPiecePositioned(Square.c3, Piece.KNIGHT_BLACK), PiecePositioned.getPosition(Square.b4));
@@ -223,10 +223,10 @@ public class MinMaxPruningBlackTest {
         Game rootGame = setupGame(Color.BLACK, GameStatus.NO_CHECK);
 
         Game childGame1 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame1), eq(2), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1);
+        when(quiescence.maximize(eq(childGame1), eq(2), eq(GameEvaluator.INFINITE_NEGATIVE), eq(GameEvaluator.INFINITE_POSITIVE))).thenReturn(1L);
 
         Game childGame2 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
-        when(quiescence.maximize(eq(childGame2), eq(2), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn(GameEvaluator.BLACK_WON);
+        when(quiescence.maximize(eq(childGame2), eq(2), eq(GameEvaluator.INFINITE_NEGATIVE), eq(1))).thenReturn((long)GameEvaluator.BLACK_WON);
 
         Game childGame3 = setupGame(Color.WHITE, GameStatus.NO_CHECK);
 
@@ -236,7 +236,7 @@ public class MinMaxPruningBlackTest {
         linkMovesToGames(rootGame, new Move[]{move1, move2, move3}, new Game[]{childGame1, childGame2, childGame3});
 
         alphaBeta.init(rootGame, new SearchContext(2));
-        int minValue = alphaBeta.minimize(rootGame, 1, GameEvaluator.INFINITE_NEGATIVE, GameEvaluator.INFINITE_POSITIVE);
+        long minValue = alphaBeta.minimize(rootGame, 1, GameEvaluator.INFINITE_NEGATIVE, GameEvaluator.INFINITE_POSITIVE);
 
         assertEquals(GameEvaluator.BLACK_WON, minValue);
